backend/seguridad_servicio/requirements-test-gestor-usuarios.txt
blinker==1.9.0
click==8.1.8
coverage==7.7.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pip==24.3.1
pluggy==1.5.0
pytest==8.3.5
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
PyJWT==2.6.0
flask-cors==5.0.1